Interlude 2 with all components
Uploaded as a zip file because there are too many files to do otherwise! Github has a 100 file upload limit, and the nature of the project defied that limit. ...It also doesn't like files bigger than 25MB, so I've chopped out the full day 1 version of the code in favor of just the day 2 version and the single file I edited on day 1 on its own.